visual speed display

+++ Core/Src/Bicycle/State.h
@@ -4,10 +4,10 @@ struct State {
     /**
      * Speed in km/h.
      */
-    int speed = 0;
+    float speed = 0;
 
     /**
      * Overall distance on this session.
      */
-    int distance = 0;
+    float distance = 0;
 };
